Remove nodecontroller from required components
Problem is that SearchCerts tries to find all known certs
in default namespace. But node controller lives in sepatate
namespace, moreover we don't have inforamation about TPO.

This change is not propoer solution for SearchCerts. But overall we need to
rethink SearchCerts one more time, looks like certtificatetpr
is not the proper place for such specific busness logic.